Fix 1 for task 12

+++ 0x00-MySQL_Advanced/100-average_weighted_score.sql
@@ -5,7 +5,7 @@ DELIMITER $$
 CREATE PROCEDURE ComputeAverageWeightedScoreForUsers (IN user_id INT)
 -- Edit Procedure body below
 BEGIN
-	DECLARE total_score FLOAT DEFAULT 0.0;
+	DECLARE total_score INT DEFAULT 0;
 	DECLARE total_weight INT DEFAULT 0;
 
 	SELECT SUM(corrections.score * projects.weight) INTO total_score,
